Move mapping intersection into the index package so it can reuse unexported helpers.
Export IntersectMappings for the indexmappings Read path without adding a new public seam for dynamicTemplatesByName.
